// src/app/api/model/invoice.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Payment } from './payment';
import { Organization } from './organization';
import { Vendor } from './vendor';


export interface Invoice { 
    id?: string;
    orgId?: string | null;
    vendorId?: string | null;
    invoiceNumber?: string | null;
    dateIssued?: string | null;
    dueDate?: string | null;
    totalAmount?: number | null;
    currency?: string | null;
    status?: string | null;
    category?: string | null;
    ocrText?: string | null;
    aiSummary?: string | null;
    anomalyScore?: number | null;
    createdAt?: string | null;
    org?: Organization;
    payments?: Array<Payment> | null;
    vendor?: Vendor;
}